backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
python-dotenv==1.0.0
openai==0.27.8
pdfplumber==0.9.0
python-docx==0.8.11
gunicorn==20.1.0
psycopg2-binary==2.9.6